+++ NEWS.md
@@ -1,6 +1,10 @@
 # rollama (development version)
 
-# rollama 0.2.2
+* query now supports logprobs output
+* update parameters in embed_text()
+* update parameters and output in show_model()
+* adds list_running_models()
+* adds option to cache responses
 
 # rollama 0.2.1